+++ src/features/island/buildings/components/building/market/lib/makeBulkBuyAmount.ts
